Shift RecyclerView to HomeFragment
