+++ app/src/main/java/udacity/designvilla/HomeActivity.java
@@ -12,11 +12,8 @@ import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.Toast;
 
 import com.example.android.splashscreenjava.R;
@@ -24,22 +21,13 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.rom4ek.arcnavigationview.ArcNavigationView;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
-
-import udacity.designvilla.adapter_view.Adapter;
-import udacity.designvilla.adapter_view.TemplateHolder;
 
 public class HomeActivity extends AppCompatActivity {
 
     private DrawerLayout mDrawerLayout;
     private ArcNavigationView mNavigationView;
     private Toolbar mToolbar;
-    private RecyclerView mRecyclerView;
-    private RecyclerView.Adapter mAdapter;
-    private RecyclerView.LayoutManager mLayoutManager;
-    private List<TemplateHolder> templateHolders;
     private FirebaseUser mUser;
 
     @Override
@@ -61,7 +49,8 @@ public class HomeActivity extends AppCompatActivity {
 
 
         FragmentTransaction tx = getFragmentManager().beginTransaction();
-        tx.replace(R.id.main_layout, new HomeFragment());
+        //Replacing just an unused LinearLayout so as to use Navigation Drawer and ToolBar in every fragment
+        tx.replace(R.id.activity_home_replaceable_view, new HomeFragment());
         tx.commit();
 
 
@@ -102,51 +91,13 @@ public class HomeActivity extends AppCompatActivity {
 
                 FragmentManager fragmentManager = getFragmentManager();
                 FragmentTransaction transaction = fragmentManager.beginTransaction();
-                transaction.replace(R.id.main_layout, fragment);
+                transaction.replace(R.id.activity_home_replaceable_view, fragment);
                 transaction.commit();
                 //close the navigation bar when clicked
                 mDrawerLayout.closeDrawers();
                 return true;
             }
         });
-
-        //RecyclerView
-        mRecyclerView = findViewById(R.id.template_recycler_view);
-
-        //Use a Staggered Layout Manager
-        mLayoutManager = new StaggeredGridLayoutManager(2,
-                StaggeredGridLayoutManager.VERTICAL);
-        mRecyclerView.setLayoutManager(mLayoutManager);
-
-        //Set Adapter
-        mAdapter = new Adapter(passDummyData());
-        mRecyclerView.setAdapter(mAdapter);
-    }
-
-    private List<TemplateHolder> passDummyData() {
-        templateHolders = new ArrayList<>();
-        int n = 1;
-        for (int i = 0; i < 15; i++) {
-            switch (n) {
-                case 1:
-                    templateHolders.add(new TemplateHolder(getDrawable(R.drawable.a1)));
-                    n = 2;
-                    break;
-                case 2:
-                    templateHolders.add(new TemplateHolder(getDrawable(R.drawable.a2)));
-                    n = 3;
-                    break;
-                case 3:
-                    templateHolders.add(new TemplateHolder(getDrawable(R.drawable.a3)));
-                    n = 4;
-                    break;
-                case 4:
-                    templateHolders.add(new TemplateHolder(getDrawable(R.drawable.a4)));
-                    n = 1;
-                    break;
-            }
-        }
-        return templateHolders;
     }
 
     @Override
